Finalized shared_pointers.  All tests pass.
I should note that the *crementReference functions in ControlBlock had to be "un-inlined".
Specifically the decrementReference function was so large that -Winline was throwing
warnings on build and rather than adjust params I figured to take it with a grain of salt
and trust its ruling of "function too large".  So I just removed inlines to those calls.