Converted Vector to use std::valarray
I'm not actually sure how I feel about this...  I'm not going to worry
about it much right now, but after the class is over I'll probably
decide on one way of doing things.  Valarray will probably be out
since it allocates things on the heap and I'd like the vector objects to
contain their members.